Don't try and skip environment re-creation
Tox doesn't pass enough info onto us from what I can see to distinguish
between times we should skip this and times we shouldn't.

There are lots of reasons why this can end-up getting triggered. The one we
were attempting to catch was when the requirements are detected as changed
by tox, in which case the environment does not need to be recreated.

However this is identical from what I can see to the situation where the
version of Python has changed... which is very relevant and we absolutely
can't skip recreation in that case.